Replace SynchronizedResolverSpec with SynchronizedResolverTests.
